requirements.txt
flask>=2.3.0
boto3>=1.28.0
python-dotenv>=1.0.0
Pillow>=10.0.0
botocore>=1.31.0
requests>=2.31.0
